Move exceptions in a separate package, resulting bugfix
