+++ flink-jpmml-scala/src/main/scala/io/radicalbit/flink/pmml/scala/api/pipeline/Pipeline.scala
@@ -19,8 +19,8 @@
 
 package io.radicalbit.flink.pmml.scala.api.pipeline
 
-import io.radicalbit.flink.pmml.scala.InputPreparationException
 import io.radicalbit.flink.pmml.scala.api._
+import io.radicalbit.flink.pmml.scala.api.exceptions.InputPreparationException
 import io.radicalbit.flink.pmml.scala.models.Prediction
 import org.apache.flink.ml.math.Vector
 import org.dmg.pmml.FieldName
